Updated Home & Login, First draft of Create New Character
Updated Home & Login pages, colors and other design elements.  Started the first of the Create a New Character pages.  Submenu is experimental.  Note new image choice for the Create New Character page - intend to use the same image for each page as create a new character.  Also intend to add a progress bar on pages 2 - end of creation of character.